refactor: organize remote datasource package structure
