+++ remote_datasource/src/main/java/com/baghdad/remoteDataSource/RemoteSearchDataSourceImpl.kt
@@ -1,9 +1,10 @@
-package com.baghdad.remoteDataSource.util
+package com.baghdad.remoteDataSource
 
-import com.baghdad.remoteDataSource.util.entity.SearchParameter
-import com.baghdad.remoteDataSource.util.entity.SearchResponse
-import com.baghdad.remoteDataSource.util.entity.toParams
-import com.baghdad.remoteDataSource.util.mapper.toDto
+import com.baghdad.remoteDataSource.request.SearchParameter
+import com.baghdad.remoteDataSource.response.SearchResponse
+import com.baghdad.remoteDataSource.request.toParams
+import com.baghdad.remoteDataSource.mapper.toDto
+import com.baghdad.remoteDataSource.util.handleRequest
 import com.baghdad.repository.datasource.remote.RemoteSearchDataSource
 import com.baghdad.repository.model.GenreDto
 import com.baghdad.repository.model.SearchResultDto
